Add files via upload
This python project is on Ramen Ratings. The dataset is from Kaggle. Just to explore fun things on this dataset I have performed EDA and found out a few interesting information regarding this dataset.
-------------------------------------------------------------------------------------------------------------------------------------------------------
The entire data is collected by a single person whose hobby is to taste different Ramen and rate them. He has tasted various Ramens available in almost all parts of the world.
The data had to be cleaned before any EDA was performed.
Following are the things noticed from the columns present in the data:
I have dropped the column Top Ten as it had only 41 filled values. This column only has the top ten ramen each year, from 2012 to 2016. and my guess is maybe it's the general top ten.
For the column Stars,  there are some text data in it, marked as Unrated. I have replaced unrated with 0 assuming he was not satisfied with those products. 

Being a student and wanting to play around the dataset I wanted to check what would this data do if I trained them and asked them to predict results using the Machine Learning algorithm. 
Since the value count of style is 7 and it's not easy to balance the data. So for the sake of practicing Machine learning, I have grouped them into Packs as 0 and others as 1. So I take the Style as my target variable and now send this data to the models.
In the Supervised Learning it is found out that 
-> Random Forest has higher accuracy
By using ensemble techniques:
-> Bagging gives higher accuracy for this dataset
I have also performed Unsupervised classification for this dataset by removing the target variable.
KMeans gave 15 percent accuracy and attaching the Kmeans predicted values to the dataset we again perform the Supervised Learning
Decision Tree with four clusters gave the highest accuracy.
Similarly, on performing Hierarchical Clustering, it gave 53 percent accuracy which is better than KMeans clustering. The same predicted values I have performed Supervised Learning again and found out that KNN gave higher accuracy.
On performing PCA in this dataset, PCA without scaling gave:
We have taken n_components as 2
With this RF has higher accuracy
On performing PCA with scaling:
We have taken n_components as 2
With this RF has higher accuracy